rename refactored utils file
