+++ practice/cli.py
@@ -19,8 +19,8 @@ sys.path.insert(0, str(project_root))
 from practice.base.context import ContextBase  # noqa: E402
 from practice.utils.cli_utils import get_utc_time_str, load_config_module  # noqa: E402
 from practice.utils.hub_utils import push_to_hub_generic  # noqa: E402
-from practice.utils.play_utils_new import play_and_generate_video_generic  # noqa: E402
-from practice.utils.train_utils_new import train_and_evaluate_network  # noqa: E402
+from practice.utils.play_utils import play_and_generate_video_generic  # noqa: E402
+from practice.utils.train_utils import train_and_evaluate_network  # noqa: E402
 
 
 def _create_parser() -> argparse.ArgumentParser:
