+++ practice/exercise4_curiosity/tests/test_curiosity_train.py
@@ -43,7 +43,7 @@ from practice.exercise4_curiosity.enhanced_reinforce import (
     EnhancedReinforceTrainer,
 )
 from practice.utils.env_utils import get_device
-from practice.utils.train_utils_new import train_and_evaluate_network
+from practice.utils.train_utils import train_and_evaluate_network
 from practice.utils_for_coding.baseline_utils import ConstantBaseline
 from practice.utils_for_coding.scheduler_utils import ConstantSchedule, ExponentialSchedule
