+++ practice/exercise7_ppo/tests/test_ppo_train.py
@@ -7,7 +7,7 @@ import pytest
 
 from practice.exercise7_ppo.config_cartpole import generate_context, get_app_config
 from practice.exercise7_ppo.ppo_exercise import PPOConfig
-from practice.utils.train_utils_new import train_and_evaluate_network
+from practice.utils.train_utils import train_and_evaluate_network
 
 
 @pytest.fixture
